Create `NumericalFormatter` (#1007)
* Add NumericalFormatter + tests

* Integrate NumericalFormatter into DataProcessor

* Update `representation` values to match the ones from RDT's `computer_representation`

* Add tests + update logic

* Typo fix

* Fix test mock bug

* Fix minor column

* Fix rebase + format_data takes a Series

* Remove test added by accident

+++ tests/integration/metadata/test_single_table.py
@@ -188,7 +188,7 @@ def test_upgrade_metadata():
             },
             'salary': {
                 'sdtype': 'numerical',
-                'representation': 'int64'
+                'representation': 'Int64'
             },
             'duration': {
                 'sdtype': 'categorical'
@@ -198,7 +198,7 @@ def test_upgrade_metadata():
             },
             'high_perc': {
                 'sdtype': 'numerical',
-                'representation': 'float64'
+                'representation': 'Float'
             },
             'placed': {
                 'sdtype': 'boolean'
